[Feat] 제대로 된 로또 한장 만들기
로또에 중복이 있을 경우와 중복이 없는 경우 오름차순으로 정리 된 제대로 된 예쁜 로또를 만들 수 있다.
